[Fix] [Fix] 관리자 대시보드 페이지 컴포넌트화 (#64)
