+++ package.json
@@ -16,7 +16,7 @@
     "jwt-decode": "^4.0.0",
     "vue": "^3.2.13",
     "vue-chartjs": "^5.3.0",
-    "vue-router": "^4.3.0",
+    "vue-router": "^4.3.2",
     "vuex": "^4.1.0"
   },
   "devDependencies": {
