+++ src/components/ComplaintDashBoardComponent.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    ComplaintDashBoardComponent
+  </div>
+</template>
+
+<script>
+export default {
+  setup () {
+    
+
+    return {}
+  }
+}
+</script>
+
+<style lang="scss" scoped>
+
+</style>
