+++ src/components/EventDashBoardComponent.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    EventDashBoardComponent
+  </div>
+</template>
+
+<script>
+export default {
+  setup () {
+    
+
+    return {}
+  }
+}
+</script>
+
+<style lang="scss" scoped>
+
+</style>
